changes inputs to float and IReadOnlyList

+++ src/Signals/Sinks/LevelMeterSignal.cs
@@ -10,14 +10,14 @@ namespace VL.Audio
             InputSignal.Value = input;
         }
         
-        public double Max;
+        public float Max;
 
         protected override void FillBuffer(float[] buffer, int offset, int count)
         {
             if (InputSignal.Value != null) 
             {
                 InputSignal.Read(buffer, offset, count);
-                var max = 0.0;
+                var max = 0.0f;
                 for (int i = offset; i < count; i++) 
                 {
                     max = Math.Max(max, Math.Abs(buffer[i]));
